+++ editor/src/components/inspector/sections/component-section/component-section.tsx
@@ -99,6 +99,8 @@ import {
 } from '../../../../core/shared/jsx-attributes'
 import type { VariableData } from '../../../canvas/ui-jsx-canvas'
 import { array } from 'prop-types'
+import { useVariablesInScopeForSelectedElement } from './variables-in-scope-utils'
+import { DataPickerPopup } from './data-picker-popup'
 
 export const VariableFromScopeOptionTestId = (idx: number) => `variable-from-scope-${idx}`
 export const DataPickerPopupButtonTestId = `data-picker-popup-button-test-id`
@@ -386,252 +388,6 @@ function getSectionHeightFromPropControl(
     return accumulatedHeight
   }
 }
-
-function usePropertyControlDescriptions(): Array<ControlDescription> {
-  return useGetPropertyControlsForSelectedComponents().flatMap((controls) =>
-    Object.values(controls.controls),
-  )
-}
-
-function arrayShapesMatch(left: Array<unknown>, right: Array<unknown>): boolean {
-  if (left.length === 0 || right.length === 0) {
-    return true
-  }
-
-  return variableShapesMatch(left[0], right[0])
-}
-
-function objectShapesMatch(left: object, right: object): boolean {
-  const keysFromLeft = Object.keys(left)
-  const keysFromRight = Object.keys(right)
-  const keysMatch =
-    keysFromLeft.length === keysFromRight.length &&
-    keysFromLeft.every((key) => keysFromRight.includes(key))
-
-  if (!keysMatch) {
-    return false
-  }
-
-  return keysFromLeft.every((key) => variableShapesMatch((left as any)[key], (right as any)[key]))
-}
-
-function variableShapesMatch(left: unknown, right: unknown): boolean {
-  if (Array.isArray(left) && Array.isArray(right)) {
-    return arrayShapesMatch(left, right)
-  }
-
-  if (typeof left === 'object' && typeof right === 'object' && left != null && right != null) {
-    return objectShapesMatch(left, right)
-  }
-
-  return typeof left === typeof right
-}
-
-function variableMatchesArrayControlDescription(
-  variable: Array<unknown>,
-  controlDescription: ArrayControlDescription,
-): boolean {
-  if (variable.length === 0) {
-    return true
-  }
-
-  return variableMatchesControlDescription(variable[0], controlDescription.propertyControl)
-}
-
-function variableMatchesObjectControlDescription(
-  variable: object,
-  controlDescription: ObjectControlDescription,
-): boolean {
-  return Object.entries(controlDescription.object).every(([key, control]) =>
-    variableMatchesControlDescription((variable as any)[key], control),
-  )
-}
-
-function variableMatchesControlDescription(
-  variable: unknown,
-  controlDescription: ControlDescription,
-): boolean {
-  const matches =
-    (typeof variable === 'string' && controlDescription.control === 'string-input') ||
-    (typeof variable === 'number' && controlDescription.control === 'number-input') ||
-    (Array.isArray(variable) &&
-      controlDescription.control === 'array' &&
-      variableMatchesArrayControlDescription(variable, controlDescription)) ||
-    (typeof variable === 'object' &&
-      variable != null &&
-      controlDescription.control === 'object' &&
-      variableMatchesObjectControlDescription(variable, controlDescription))
-
-  return matches
-}
-
-type PropertyValue = { type: 'existing'; value: unknown } | { type: 'not-found' }
-
-function usePropertyValue(selectedView: ElementPath, propertyPath: PropertyPath): PropertyValue {
-  const allElementProps = useEditorState(
-    Substores.metadata,
-    (store) => store.editor.allElementProps,
-    'usePropertyValue allElementProps',
-  )
-  const propsForThisElement = allElementProps[EP.toString(selectedView)] ?? null
-  if (propsForThisElement == null) {
-    return { type: 'not-found' }
-  }
-
-  const prop = propsForThisElement[propertyPath.propertyElements[0]] ?? null
-  if (prop == null) {
-    return { type: 'not-found' }
-  }
-
-  return { type: 'existing', value: prop }
-}
-
-interface DataPickerPopupProps {
-  closePopup: () => void
-  style: React.CSSProperties
-  propPath: PropertyPath
-}
-
-const DataPickerPopup = React.memo(
-  React.forwardRef<HTMLDivElement, DataPickerPopupProps>((props, forwardedRef) => {
-    const { closePopup, propPath } = props
-
-    const selectedViewPathRef = useRefEditorState(
-      (store) => store.editor.selectedViews.at(0) ?? null,
-    )
-
-    const colorTheme = useColorTheme()
-    const dispatch = useDispatch()
-
-    const onTweakProperty = React.useCallback(
-      (name: string, definedElsewhere: string | null) => (e: React.MouseEvent) => {
-        if (selectedViewPathRef.current == null) {
-          return
-        }
-
-        e.stopPropagation()
-        e.preventDefault()
-
-        const definedElseWhereArray = optionalMap((d) => [d], definedElsewhere) ?? []
-
-        dispatch([
-          setProp_UNSAFE(
-            selectedViewPathRef.current,
-            propPath,
-            jsExpressionOtherJavaScriptSimple(name, definedElseWhereArray),
-          ),
-        ])
-      },
-      [dispatch, propPath, selectedViewPathRef],
-    )
-
-    const variableNamesInScope = useVariablesInScopeForSelectedElement(
-      selectedViewPathRef.current ?? EP.emptyElementPath,
-      props.propPath,
-    )
-
-    return (
-      <div
-        style={{
-          background: 'transparent',
-          position: 'fixed',
-          top: 0,
-          left: 0,
-          right: 0,
-          bottom: 0,
-          zIndex: 1, // so it's above the inspector
-        }}
-        onClick={closePopup}
-      >
-        <FlexColumn
-          ref={forwardedRef}
-          tabIndex={0}
-          style={{
-            ...props.style,
-            backgroundColor: colorTheme.neutralBackground.value,
-            padding: '8px 16px',
-            boxShadow: UtopiaStyles.shadowStyles.mid.boxShadow,
-            borderRadius: UtopiaTheme.inputBorderRadius,
-            alignItems: 'flex-start',
-          }}
-          data-testid={DataPickerPopupTestId}
-        >
-          <div style={{ fontSize: 14, fontWeight: 400, marginBottom: 16 }}>
-            <span>Data</span>
-          </div>
-          {variableNamesInScope.map(
-            ({ variableName, definedElsewhere, value, displayName, depth = 0 }, idx) => {
-              return (
-                <Button
-                  data-testid={VariableFromScopeOptionTestId(idx)}
-                  key={variableName}
-                  onClick={onTweakProperty(variableName, definedElsewhere)}
-                  style={{ width: '100%' }}
-                >
-                  <UIGridRow
-                    padded={false}
-                    variant='<--1fr--><--1fr-->'
-                    style={{
-                      justifyContent: 'space-between',
-                      alignItems: 'flex-start',
-                      gap: 8,
-                      width: '100%',
-                    }}
-                  >
-                    <div>
-                      <span
-                        style={{
-                          marginLeft: 4 * depth,
-                          borderRadius: 2,
-                          fontWeight: 400,
-                        }}
-                      >
-                        {depth > 0 ? (
-                          <span
-                            style={{
-                              borderLeft: `1px solid ${colorTheme.neutralBorder.value}`,
-                              borderBottom: `1px solid ${colorTheme.neutralBorder.value}`,
-                              width: 5,
-                              display: 'inline-block',
-                              height: 9,
-                              marginRight: 4,
-                              position: 'relative',
-                              top: -2,
-                              marginLeft: (depth - 1) * 8,
-                            }}
-                          ></span>
-                        ) : null}
-                        {displayName}
-                      </span>
-                    </div>
-                    <div
-                      style={{
-                        display: 'flex',
-                        justifyContent: 'flex-end',
-                      }}
-                    >
-                      <span
-                        style={{
-                          fontWeight: 400,
-                          color: colorTheme.neutralForeground.value,
-                          textOverflow: 'ellipsis',
-                          maxWidth: 130,
-                          overflow: 'hidden',
-                        }}
-                      >
-                        {value}
-                      </span>
-                    </div>
-                  </UIGridRow>
-                </Button>
-              )
-            },
-          )}
-        </FlexColumn>
-      </div>
-    )
-  }),
-)
 
 interface RowForArrayControlProps extends AbstractRowForControlProps {
   controlDescription: ArrayControlDescription
@@ -1311,168 +1067,4 @@ export class ComponentSection extends React.Component<
       return <ComponentSectionInner {...this.props} />
     }
   }
-}
-
-interface VariableOption {
-  variableName: string
-  displayName: string
-  definedElsewhere: string | null
-  value: string
-  depth?: number
-}
-
-function valuesFromObject(
-  name: string,
-  objectName: string,
-  value: object | null,
-  depth: number,
-  displayName: string,
-): Array<VariableOption> {
-  if (value == null) {
-    return [{ displayName: displayName, variableName: name, definedElsewhere: null, value: `null` }]
-  }
-
-  const patchDefinedElsewhereInfo = (variable: VariableOption) => ({
-    variableName: variable.variableName,
-    value: variable.value,
-    definedElsewhere: objectName,
-    displayName: variable.displayName,
-    depth: variable.depth,
-  })
-
-  if (Array.isArray(value)) {
-    return [
-      patchDefinedElsewhereInfo({
-        displayName: displayName,
-        variableName: name,
-        definedElsewhere: objectName,
-        value: `[ ]`,
-        depth: depth,
-      }),
-    ].concat(
-      value.flatMap((v, idx) =>
-        valuesFromVariable(`${name}[${idx}]`, v, depth + 1, `${displayName}[${idx}]`).map(
-          (variable) => patchDefinedElsewhereInfo(variable),
-        ),
-      ),
-    )
-  }
-
-  return [
-    patchDefinedElsewhereInfo({
-      displayName: displayName,
-      variableName: name,
-      definedElsewhere: objectName,
-      value: `{ }`,
-      depth: depth,
-    }),
-  ].concat(
-    Object.entries(value).flatMap(([key, field]) =>
-      valuesFromVariable(`${name}['${key}']`, field, depth + 1, key).map((variable) =>
-        patchDefinedElsewhereInfo(variable),
-      ),
-    ),
-  )
-}
-
-function valuesFromVariable(
-  name: string,
-  value: unknown,
-  depth: number,
-  displayName: string,
-): Array<VariableOption> {
-  switch (typeof value) {
-    case 'bigint':
-    case 'boolean':
-    case 'number':
-    case 'string':
-    case 'undefined':
-      return [
-        {
-          displayName: displayName,
-          variableName: name,
-          definedElsewhere: name,
-          value: `${value}`,
-          depth: depth,
-        },
-      ]
-    case 'object':
-      return valuesFromObject(name, name, value, depth, displayName)
-    case 'function':
-    case 'symbol':
-      return []
-  }
-}
-
-function orderVariablesInScope(
-  variableNamesInScope: VariableData,
-  controlDescriptions: Array<ControlDescription>,
-  currentPropertyValue: PropertyValue,
-): Array<[string, unknown]> {
-  let valuesMatchingPropertyDescription: [string, unknown][] = []
-  let valuesMatchingPropertyShape: [string, unknown][] = []
-  let restOfValues: [string, unknown][] = []
-
-  for (const [name, { spiedValue }] of Object.entries(variableNamesInScope)) {
-    const valueMatchesControlDescription = controlDescriptions.some((d) =>
-      variableMatchesControlDescription(spiedValue, d),
-    )
-    const valueMatchesCurrentPropValue =
-      currentPropertyValue.type === 'existing' &&
-      variableShapesMatch(currentPropertyValue.value, spiedValue)
-
-    if (valueMatchesControlDescription) {
-      valuesMatchingPropertyDescription.push([name, spiedValue])
-    } else if (valueMatchesCurrentPropValue) {
-      valuesMatchingPropertyShape.push([name, spiedValue])
-    } else {
-      restOfValues.push([name, spiedValue])
-    }
-  }
-
-  return [...valuesMatchingPropertyDescription, ...valuesMatchingPropertyShape, ...restOfValues]
-}
-
-function useVariablesInScopeForSelectedElement(
-  selectedView: ElementPath,
-  propertyPath: PropertyPath,
-): Array<VariableOption> {
-  const selectedViewPath = useEditorState(
-    Substores.selectedViews,
-    (store) => store.editor.selectedViews.at(0) ?? null,
-    'useVariablesInScopeForSelectedElement selectedViewPath',
-  )
-
-  const variablesInScope = useEditorState(
-    Substores.restOfEditor,
-    (store) => store.editor.variablesInScope,
-    'useVariablesInScopeForSelectedElement variablesInScope',
-  )
-
-  const controlDescriptions = usePropertyControlDescriptions()
-  const currentPropertyValue = usePropertyValue(selectedView, propertyPath)
-
-  const variableNamesInScope = React.useMemo((): Array<VariableOption> => {
-    if (selectedViewPath == null) {
-      return []
-    }
-
-    const variablesInScopeForSelectedPath = variablesInScope[EP.toString(selectedViewPath)]
-
-    if (variablesInScopeForSelectedPath == null) {
-      return []
-    }
-
-    const orderedVariablesInScope = orderVariablesInScope(
-      variablesInScopeForSelectedPath,
-      controlDescriptions,
-      currentPropertyValue,
-    )
-
-    return orderedVariablesInScope.flatMap(([name, variable]) =>
-      valuesFromVariable(name, variable, 0, name),
-    )
-  }, [controlDescriptions, currentPropertyValue, selectedViewPath, variablesInScope])
-
-  return variableNamesInScope
 }
